# ARROW-2558: [Plasma] avoid walk through all the objects when a client disconnects