@pcmoritz 2018-05-15
Performance impact:

Without this PR:
```
[ 78.57%] ··· Running plasma.SimplePlasmaLatency.time_plasma_put                                                                                                             498.27ms
[ 85.71%] ··· Running plasma.SimplePlasmaLatency.time_plasma_putget                                                                                                          694.21ms
[ 92.86%] ··· Running plasma.SimplePlasmaThroughput.time_plasma_put_data                                                                                                           ok
[ 92.86%] ···· 
               ========== ==========
                 param1             
               ---------- ----------
                  1000     552.60μs 
                 100000    768.20μs 
                10000000    5.61ms  
               ========== ==========
```

With this PR:
```
[ 78.57%] ··· Running plasma.SimplePlasmaLatency.time_plasma_put                                                                                                             500.77ms
[ 85.71%] ··· Running plasma.SimplePlasmaLatency.time_plasma_putget                                                                                                          697.12ms
[ 92.86%] ··· Running plasma.SimplePlasmaThroughput.time_plasma_put_data                                                                                                           ok
[ 92.86%] ···· 
               ========== ==========
                 param1             
               ---------- ----------
                  1000     557.60μs 
                 100000    741.20μs 
                10000000    4.89ms  
               ========== ==========
```

So looks like the performance is approximately the same.